Rename collections to records
