+++ src/app/progress/Progress.tsx
@@ -23,11 +23,11 @@ import React, { useState } from 'react';
 import Hammer from 'react-hammerjs';
 import { connect } from 'react-redux';
 import { DestinyAccount } from '../accounts/destiny-account';
-import '../collections/PresentationNode.scss';
 import { D2ManifestDefinitions } from '../destiny2/d2-definitions';
 import CollapsibleTitle from '../dim-ui/CollapsibleTitle';
 import ErrorBoundary from '../dim-ui/ErrorBoundary';
 import { InventoryBuckets } from '../inventory/inventory-buckets';
+import '../records/PresentationNode.scss';
 import Milestones from './Milestones';
 import './progress.scss';
 import Pursuits from './Pursuits';
