+++ src/app/records/Collectible.tsx
