+++ src/app/records/Metric.m.scss
