+++ src/app/records/Metric.tsx
