+++ src/app/records/MetricBanner.m.scss
