+++ src/app/records/MetricBanner.tsx
