+++ src/app/records/Record.tsx
