+++ src/app/records/collections.scss
